feat: install react-router-dom and add it
