+++ src/components/NotFound/NotFound.tsx
@@ -1,0 +1,13 @@
+import { Link } from 'react-router-dom';
+
+const NotFound = () => {
+  return (
+    <div style={{ textAlign: 'center', padding: '4rem' }}>
+      <h1>404 - Page Not Found</h1>
+      <p>The page you are looking for does not exist.</p>
+      <Link to="/">Go back home</Link>
+    </div>
+  );
+};
+
+export default NotFound;
